Rework result module extraction base class to be more ergonomic
